# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: CIDB Tender API
  version: 1.0.0
  description: "Standalone normalized API over publicly available CIDB tender
    information. Authenticate with the `X-API-Key` header. Source: Construction
    Industry Development Board (CIDB)."
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: X-API-Key
      in: header
      description: API key issued for this service
...
paths:
...
  /admin/sync:
    post:
      tags:
        - admin
      description: Trigger a CIDB synchronization. Runs asynchronously; poll GET
        /admin/sync/:id for progress.
      security:
        - apiKey: []
      responses:
        "202":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  syncId:
                    type: string
                required:
                  - status
                  - syncId
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
  /admin/sync/history:
    get:
      tags:
        - admin
      description: Recent synchronization runs, newest first.
      parameters:
        - schema:
            type: integer
            minimum: 1
            default: 1
          in: query
          name: page
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          in: query
          name: limit
          required: false
        - schema:
            type: string
            enum:
              - RUNNING
              - COMPLETED
              - PARTIAL
              - FAILED
          in: query
          name: status
          required: false
      security:
        - apiKey: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        source:
                          type: string
                        status:
                          type: string
                        startedAt:
                          type: string
                        completedAt:
                          type: string
                          nullable: true
                        recordsDiscovered:
                          type: number
                        recordsCreated:
                          type: number
                        recordsUpdated:
                          type: number
                        recordsUnchanged:
                          type: number
                        recordsFailed:
                          type: number
                        documentsFound:
                          type: number
                        errorCount:
                          type: number
                        durationMs:
                          type: number
                          nullable: true
                        errorMessage:
                          type: string
                          nullable: true
                      required:
                        - id
                        - source
                        - status
                        - startedAt
                        - completedAt
                        - recordsDiscovered
                        - recordsCreated
                        - recordsUpdated
                        - recordsUnchanged
                        - recordsFailed
                        - documentsFound
                        - errorCount
                        - durationMs
                        - errorMessage
                      additionalProperties: false
                  pagination:
                    type: object
                    properties:
                      page:
                        type: number
                      limit:
                        type: number
                      total:
                        type: number
                      totalPages:
                        type: number
                    required:
                      - page
                      - limit
                      - total
                      - totalPages
                    additionalProperties: false
                required:
                  - data
                  - pagination
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
  /admin/sync/{id}:
    get:
      tags:
        - admin
      description: Status and counters for one synchronization run, including its most
        recent errors.
      parameters:
        - schema:
            type: string
            minLength: 1
            maxLength: 100
          in: path
          name: id
          required: true
      security:
        - apiKey: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  source:
                    type: string
                  status:
                    type: string
                  startedAt:
                    type: string
                  completedAt:
                    type: string
                    nullable: true
                  recordsDiscovered:
                    type: number
                  recordsCreated:
                    type: number
                  recordsUpdated:
                    type: number
                  recordsUnchanged:
                    type: number
                  recordsFailed:
                    type: number
                  documentsFound:
                    type: number
                  errorCount:
                    type: number
                  durationMs:
                    type: number
                    nullable: true
                  errorMessage:
                    type: string
                    nullable: true
                  recentErrors:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        syncRunId:
                          type: string
                        externalId:
                          type: string
                          nullable: true
                        url:
                          type: string
                          nullable: true
                        stage:
                          type: string
                        errorType:
                          type: string
                        message:
                          type: string
                        retryCount:
                          type: number
                        createdAt:
                          type: string
                      required:
                        - id
                        - syncRunId
                        - externalId
                        - url
                        - stage
                        - errorType
                        - message
                        - retryCount
                        - createdAt
                      additionalProperties: false
                required:
                  - id
                  - source
                  - status
                  - startedAt
                  - completedAt
                  - recordsDiscovered
                  - recordsCreated
                  - recordsUpdated
                  - recordsUnchanged
                  - recordsFailed
                  - documentsFound
                  - errorCount
                  - durationMs
                  - errorMessage
                  - recentErrors
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
  /admin/errors:
    get:
      tags:
        - admin
      description: Recent ingestion errors across sync runs.
      parameters:
        - schema:
            type: integer
            minimum: 1
            default: 1
          in: query
          name: page
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
          in: query
          name: limit
          required: false
        - schema:
            type: string
          in: query
          name: syncRunId
          required: false
        - schema:
            type: string
            maxLength: 100
          in: query
          name: stage
          required: false
      security:
        - apiKey: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        syncRunId:
                          type: string
                        externalId:
                          type: string
                          nullable: true
                        url:
                          type: string
                          nullable: true
                        stage:
                          type: string
                        errorType:
                          type: string
                        message:
                          type: string
                        retryCount:
                          type: number
                        createdAt:
                          type: string
                      required:
                        - id
                        - syncRunId
                        - externalId
                        - url
                        - stage
                        - errorType
                        - message
                        - retryCount
                        - createdAt
                      additionalProperties: false
                  pagination:
                    type: object
                    properties:
                      page:
                        type: number
                      limit:
                        type: number
                      total:
                        type: number
                      totalPages:
                        type: number
                    required:
                      - page
                      - limit
                      - total
                      - totalPages
                    additionalProperties: false
                required:
                  - data
                  - pagination
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
servers:
  - url: /api/v1
    description: Current server (v1)
...